pyproject.toml
[tool.poetry]
name = "self-study-project"
version = "0.1.0"
description = ""
authors = ["dmitriy_leonov"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "4.2.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"
django-phonenumber-field = {extras = ["phonenumbers"], version = "^8.0.0"}
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.15.2"
pytz = "^2024.1"
pillow = "^10.4.0"
black = "^24.8.0"
django-filter = "^24.3"
drf-yasg = "^1.21.7"
coverage = "^7.6.1"
django-cors-headers = "^4.4.0"
setuptools = "^73.0.1"
coreapi = "^2.3.3"
pytils = "^0.4.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
